refactor(parser): extract source line mapper
